refactor in settings, update to install and refactor in key file config
