+++ settings/keys.py
@@ -65,3 +65,20 @@ KEYS_CONFIG = [
 ]
 
 
+for current in GroupConfig.GROUPS:
+    KEYS_CONFIG.extend(
+                [
+                    Key(
+                        [MOD], 
+                        GroupConfig.position(current.name), 
+                        lazy.group[current.name].toscreen(),
+                        desc="hack para cambiar al grupo dado por posiciones entre 1 y len(grupos)"
+                    ),
+                    Key(
+                        [MOD, "shift"], 
+                        GroupConfig.position(current.name), 
+                        lazy.window.togroup(current.name, switch_group=True),
+                        desc="hack para mover una ventana al grupo en el rango de 1 y len(grupos)"
+                    )
+                ]
+            )
